001--Adjust table generation logic
