+++ src/app/components/table/tile.model.ts
@@ -1,0 +1,8 @@
+import { PeriodicElement } from "../../models/periodic-element.model";
+
+export interface Tile {
+    column: number;
+    row: number;
+    element?: PeriodicElement;
+  }
+
