+++ src/app/models/periodic-element.model.ts
@@ -1,0 +1,7 @@
+export interface PeriodicElement {
+    position: number;
+    name: string;
+    weight: number;
+    symbol: string;
+    rowPosition: number;
+  }
